removeProfilePicture mutation is resolved with a common resolver

+++ server/src/graphql/Profile/updateProfilePicture.js
@@ -9,34 +9,52 @@ const PATH = 'profile/'
 
 const PROFILE_PICTUE = 'profilePicture'
 
+const UPLOAD = 'upload'
+const REMOVE = 'remove'
+
+const mainResolver = operation => async (_, { image }, { user: { id } }) => {
+	try {
+		const profileData = await Profile.findOne({ user: id }, PROFILE_PICTUE)
+
+		const currentImage = profileData[PROFILE_PICTUE]
+
+		if (!currentImage && operation === REMOVE) {
+			return sendErrorMessage('no image found')
+		}
+
+		if (currentImage) {
+			const hasDeleted = await deleteImage(currentImage)
+
+			if (hasDeleted.result !== 'ok')
+				return sendErrorMessage('something went worng')
+
+			if (operation === REMOVE) {
+				profileData[PROFILE_PICTUE] = ''
+
+				await profileData.save()
+
+				return sendMessage('Picture successfully deleted')
+			}
+		}
+
+		const imageID = await uploadImage(image, { folder: PATH, ...imageConfig })
+
+		profileData[PROFILE_PICTUE] = imageID
+
+		const res = await profileData.save()
+
+		if (res) return sendMessage('Profile Picture uploaded successfully')
+
+		return sendErrorMessage('error')
+	} catch (err) {
+		return sendErrorMessage(err)
+	}
+}
+
 const resolvers = {
 	Mutation: {
-		uploadProfilePicture: async (_, { image }, { user: { id } }) => {
-			try {
-				const profileData = await Profile.findOne({ user: id }, PROFILE_PICTUE)
-
-				const currentImage = profileData[PROFILE_PICTUE]
-
-				if (currentImage) {
-					const hasDeleted = await deleteImage(currentImage)
-
-					if (hasDeleted.result !== 'ok')
-						return sendErrorMessage('something went worng')
-				}
-
-				const imageID = await uploadImage(image, { folder: PATH, ...imageConfig })
-
-				profileData[PROFILE_PICTUE] = imageID
-
-				const res = await profileData.save()
-
-				if (res) return sendMessage('Profile Picture uploaded successfully')
-
-				return sendErrorMessage('error')
-			} catch (err) {
-				return sendErrorMessage(err)
-			}
-		},
+		uploadProfilePicture: mainResolver(UPLOAD),
+		removeProfilePicture: mainResolver(REMOVE),
 	},
 }
